+++ db/migrate/20230514003307_create_response.rb
@@ -3,7 +3,7 @@ class CreateResponse < ActiveRecord::Migration[7.0]
     create_table :responses do |t|
       t.string :response
       t.boolean :acceptance
-      t.integer :evaluation
+      t.float :evaluation
 
       t.timestamps
     end
